This commit fix follows: 1. update crate common/cctrusetd_base does not depend on std crates 2. change crate name from cctrusted to cctrusted_vm for VM SDK to in-sync with Python latest update 3. define trait CCTrustedApi in crate common/cctrusetd_base to define general cc trusted APIs 4. change api to sdk concept in cctrusted_vm crate 5. move CVM related functions to cctrusted_vm crate 6. update TCG algo list to reflact review comments by Jiewen
