+++ common/rust/cctrusted_base/Cargo.toml
@@ -12,5 +12,6 @@ anyhow = "1.0"
 base64 = "0.13.0"
 log = "0.4.20"
 sha2 = "0.10"
-nix = "0.26.2"
 lazy_static = "1.4.0"
+hashbrown = "0.14"
+
