+++ common/rust/cctrusted_base/src/binary_blob.rs
@@ -36,7 +36,7 @@ pub fn dump_data(data: &Vec<u8>) {
         match printable.iter().position(|&c| c == (v as char)) {
             Some(_) => {
                 if v < 0x9 || v > 0xD {
-                    printstr.push_str(std::str::from_utf8(&[v]).unwrap());
+                    printstr.push_str(core::str::from_utf8(&[v]).unwrap());
                 } else {
                     printstr.push_str(".");
                 }
