+++ common/rust/cctrusted_base/src/lib.rs
@@ -1,9 +1,10 @@
 #[macro_use]
 extern crate lazy_static;
 
+pub mod api;
+pub mod api_data;
 pub mod binary_blob;
 pub mod cc_type;
-pub mod tee;
 pub mod eventlog;
 pub mod tcg;
 pub mod tdx;
