+++ common/rust/cctrusted_base/src/tdx/mod.rs
@@ -2,4 +2,3 @@ pub mod common;
 pub mod quote;
 pub mod report;
 pub mod rtmr;
-pub mod tdx;
