+++ common/rust/cctrusted_base/src/tdx/rtmr.rs
@@ -1,5 +1,5 @@
 use crate::tcg::*;
-use std::collections::HashMap;
+use hashbrown::HashMap;
 
 #[allow(dead_code)]
 pub struct TdxRTMR {
