+++ common/rust/cctrusted_base/src/tpm/quote.rs
@@ -1,7 +1,5 @@
 // return of API parse_cc_report()
-pub struct TpmQuote {
-    // TODO
-}
+pub struct TpmQuote {}
 
 impl TpmQuote {
     pub fn parse_tpm_quote(_quote: Vec<u8>) -> Result<TpmQuote, anyhow::Error> {
