--- vmsdk/rust/cctrusted/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod api;
-pub mod api_data;
