+++ vmsdk/rust/cctrusted_vm/Cargo.toml
@@ -1,13 +1,15 @@
 [package]
-name = "cctrusted"
+name = "cctrusted_vm"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
-name = "cctrusted"
+name = "cctrusted_vm"
 path = "src/lib.rs"
 
 [dependencies]
 cctrusted_base = { path = "../../../common/rust/cctrusted_base" }
 anyhow = "1.0"
 log = "0.4.20"
+nix = "0.26.2"
+base64 = "0.13.0"
