+++ vmsdk/rust/sample/src/cc-sample.rs
@@ -1,5 +1,8 @@
-use cctrusted::api::*;
-use cctrusted::api_data::*;
+use cctrusted_base::api::*;
+use cctrusted_base::api_data::*;
+use cctrusted_base::cc_type::TeeType;
+use cctrusted_base::tdx::quote::TdxQuote;
+use cctrusted_vm::sdk::API;
 use log::*;
 
 fn main() {
@@ -11,7 +14,7 @@ fn main() {
 
     // retrieve cc report with API "get_cc_report"
     info!("call cc trusted API [get_cc_report] to retrieve cc report!");
-    let report = match get_cc_report(nonce, data, ExtraArgs {}) {
+    let report = match API::get_cc_report(nonce, data, ExtraArgs {}) {
         Ok(q) => q,
         Err(e) => {
             error!("error getting TDX report: {:?}", e);
@@ -21,11 +24,11 @@ fn main() {
 
     // dump the cc report with API "dump_cc_report"
     info!("call cc trusted API [dump_cc_report] to dump cc report!");
-    dump_cc_report(&report.cc_report);
+    API::dump_cc_report(&report.cc_report);
 
     // parse the cc report with API "parse_cc_report"
-    if report.cc_type == TYPE_TDX {
-        let tdx_quote: CcParsedTdxReport = match CcReport::parse_cc_report(report.cc_report) {
+    if report.cc_type == TeeType::TDX {
+        let tdx_quote: TdxQuote = match CcReport::parse_cc_report(report.cc_report) {
             Ok(q) => q,
             Err(e) => {
                 error!("error parse tdx quote: {:?}", e);
@@ -38,9 +41,9 @@ fn main() {
         );
     }
 
-    // get TEE default algorithm with API "get_default_algorithm"
-    info!("call cc trusted API [get_default_algorithm] to get TEE supported algorithm!");
-    match get_default_algorithm() {
+    // get CVM default algorithm with API "get_default_algorithm"
+    info!("call cc trusted API [get_default_algorithm] to get CVM supported algorithm!");
+    match API::get_default_algorithm() {
         Ok(algorithm) => {
             info!("supported algorithm: {}", algorithm.algo_id_str);
             ();
